rename PufferAgent
