+++ metta/agent/policy_store.py
@@ -23,7 +23,7 @@ from omegaconf import DictConfig, ListConfig
 from torch import nn
 
 from metta.agent.metta_agent import DistributedMettaAgent, MettaAgent, make_policy
-from metta.rl.pufferlib.policy import PufferAgent, load_policy
+from metta.rl.policy import MettaAgentAdapter, load_policy
 from metta.util.config import Config
 from metta.util.wandb.wandb_context import WandbRun
 
@@ -54,10 +54,10 @@ class PolicyRecord:
             self._local_path = pr.local_path()
         return self._policy
 
-    def policy_as_metta_agent(self) -> Union[MettaAgent, DistributedMettaAgent, PufferAgent]:
+    def policy_as_metta_agent(self) -> Union[MettaAgent, DistributedMettaAgent, MettaAgentAdapter]:
         """Get the policy as a MettaAgent or DistributedMettaAgent."""
         policy = self.policy()
-        if not isinstance(policy, (MettaAgent, DistributedMettaAgent, PufferAgent)):
+        if not isinstance(policy, (MettaAgent, DistributedMettaAgent, MettaAgentAdapter)):
             raise TypeError(f"Expected MettaAgent or DistributedMettaAgent, got {type(policy).__name__}")
         return policy
