+++ metta/rl/policy.py
@@ -28,11 +28,18 @@ def load_policy(path: str, device: str = "cpu", puffer: DictConfig = None):
 
     policy = instantiate(puffer, env=env, policy=None)
     policy.load_state_dict(weights)
-    policy = PufferAgent(policy).to(device)
+    policy = MettaAgentAdapter(policy).to(device)
     return policy
 
 
-class PufferAgent(nn.Module):
+class MettaAgentAdapter(nn.Module):
+    """Adapter to make loaded Puffer policies compatible with MettaAgent interface.
+
+    This adapter wraps policies loaded from checkpoints and translates their
+    outputs to match the expected MettaAgent interface, handling naming
+    differences like critic→value, hidden→logits, etc.
+    """
+
     def __init__(self, policy: nn.Module):
         super().__init__()
         self.policy = policy
@@ -47,7 +54,6 @@ class PufferAgent(nn.Module):
         """
         hidden, critic = self.policy(obs, state)  # using variable names from LSTMWrapper
         action, logprob, logits_entropy = sample_logits(hidden, action)
-        # explanation of var names in the docstring above
         return action, logprob, logits_entropy, critic, hidden
 
     def activate_actions(self, actions_names, actions_max_params, device):
